refactor: move package dfget/util to pkg/algorithm
