+++ cmd/dfdaemon/app/init.go
@@ -29,7 +29,7 @@ import (
 	"github.com/dragonflyoss/Dragonfly/dfdaemon/config"
 	"github.com/dragonflyoss/Dragonfly/dfdaemon/constant"
 	dfgetcfg "github.com/dragonflyoss/Dragonfly/dfget/config"
-	"github.com/dragonflyoss/Dragonfly/dfget/util"
+	"github.com/dragonflyoss/Dragonfly/pkg/algorithm"
 	"github.com/dragonflyoss/Dragonfly/pkg/dflog"
 	"github.com/dragonflyoss/Dragonfly/pkg/errortypes"
 	"github.com/dragonflyoss/Dragonfly/pkg/fileutils"
@@ -50,7 +50,7 @@ func adjustSupernodeList(nodes []string) []string {
 	case 1:
 		return append(nodes, nodes[0])
 	default:
-		util.Shuffle(nodesLen, func(i, j int) {
+		algorithm.Shuffle(nodesLen, func(i, j int) {
 			nodes[i], nodes[j] = nodes[j], nodes[i]
 		})
 		return append(nodes, nodes...)
