+++ pkg/algorithm/algorithm.go
@@ -16,6 +16,59 @@
 
 package algorithm
 
+import (
+	"math/rand"
+	"time"
+)
+
+func init() {
+	rand.Seed(time.Now().UnixNano())
+}
+
+// ContainsString returns whether the value is in arr.
+func ContainsString(arr []string, value string) bool {
+	for _, v := range arr {
+		if v == value {
+			return true
+		}
+	}
+	return false
+}
+
+// Shuffle pseudo-randomizes the order of elements.
+// n is the number of elements.
+// swap swaps the elements with indexes i and j.
+// copy from rand.Shuffle of go1.10.
+func Shuffle(n int, swap func(int, int)) {
+	if n < 2 {
+		return
+	}
+	i := n - 1
+	for ; i > 1<<31-1-1; i-- {
+		j := int(rand.Int63n(int64(i + 1)))
+		swap(i, j)
+	}
+	for ; i > 0; i-- {
+		j := int(int31n(int32(i + 1)))
+		swap(i, j)
+	}
+}
+
+func int31n(n int32) int32 {
+	v := rand.Uint32()
+	prod := uint64(v) * uint64(n)
+	low := uint32(prod)
+	if low < uint32(n) {
+		thresh := uint32(-n) % uint32(n)
+		for low < thresh {
+			v = rand.Uint32()
+			prod = uint64(v) * uint64(n)
+			low = uint32(prod)
+		}
+	}
+	return int32(prod >> 32)
+}
+
 // GCDSlice returns the greatest common divisor of a slice.
 // It returns 1 when s is empty because that any number divided by 1 is still
 // itself.
